+++ out/util.js
@@ -443,8 +443,15 @@ class Project {
     readonly = false;
     disableCopy = false;
     createFile(name, text, lang) {
+        let res = resolveHook(listenHooks.addFile, name);
+        if (res == 1) {
+            console.warn("Failed creation prevented from hook");
+            return;
+        }
         let f = new FFile(this, name, text, lang);
         this.files.push(f);
+        f.open();
+        return f;
     }
     hasEditor() {
         return (this.curFile?.curEditor != null);
@@ -455,9 +462,21 @@ class Project {
     init() {
         postSetupEditor(this);
     }
+    get isTutor() {
+        return this.disableCopy;
+    }
 }
 class FFile {
     constructor(p, name, text, lang) {
+        if (!lang) {
+            let ext = name.split(".").pop();
+            let map = {
+                "html": "html",
+                "css": "css",
+                "js": "javascript"
+            };
+            lang = map[ext] || "text";
+        }
         this.p = p;
         this.name = name;
         this.text = text;
@@ -506,7 +525,7 @@ class FFile {
             this.link = link;
             link.textContent = this.name;
             link.className = "file-link";
-            this.p.d_files.appendChild(link);
+            this.p.d_files.insertBefore(link, this.p.d_files.lastChild);
             this.p.openFiles.push(this);
             let t = this;
             link.onmousedown = function () {
@@ -534,11 +553,12 @@ class FFile {
                         document.activeElement.blur();
                         // is tutor editor
                         let cursor = cont.querySelector(".inputarea.monaco-mouse-cursor-text");
-                        cursor.style.backgroundColor = "white";
+                        cursor.style.backgroundColor = "var(--col)";
                         cursor.style.zIndex = "1";
                         cursor.style.width = "2px";
                         cursor.style.height = "19px";
-                        cursor.style.border = "solid 1px white";
+                        cursor.style.border = "solid 1px var(--col)";
+                        cursor.classList.add("custom-cursor");
                     }
                     // console.warn("trigger!");
                     // editor.trigger("blur","",[]);
@@ -589,4 +609,14 @@ class FFile {
         loadEditorTheme();
     }
 }
+function cursorLoop() {
+    let delay = 500;
+    let list = document.querySelectorAll(".custom-cursor");
+    for (const c of list) {
+        if (c.classList.toggle("hide"))
+            delay = 500;
+    }
+    setTimeout(cursorLoop, delay);
+}
+cursorLoop();
 //# sourceMappingURL=util.js.map