+++ styles/changelog.md
@@ -4,3 +4,14 @@
 - fixed bubbles scrolling with editor on independent editors
 - added true readonly support to tutor editor with custom cursor that always shows up
 - start of figuring out how to set up the sequence tutorial system
+
+### 1/5/24 - Claeb
+- added light theme support for forced custom cursor
+- added blinking to custom curor
+- added "add file" button to d-files list in lesson page
+- added more to LEvent (Task system)
+- added Task (Subtask system)
+- added Hook system for listening when tasks are successfully completed
+- added some real tasks and messages
+- added animations for bubbles being added and removed
+- can click okay on global bubbles and hook get's fired and proceeds when adding the correct file